Remove redundant script and merge features
